[FIX] 실시간 채팅
